Interactive client search, less debug
This exploits the generalisation to expand fuzzy search to clients (now
to see if terminal emulators really are that distinguishable).

This still could benefit from a further canonisation, in that the
`view_{client,tag}_by_substr` functions are superfluous. This could be
made explicitly part of the behaviour of
`search_{client,tag}_interactive`. And that, in turn, could be made
generic to a `search_anything_interactive` that takes a `source`
parameter.